third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGMarkerElement-svgdom-setOrientToAngle-call-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the SVGMarkerElement object by using it's 'setOrientToAngle' method

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_AUTO
PASS markerElement.orientAngle.baseVal.value is 0
PASS markerElement.setOrientToAngle() threw exception TypeError: Failed to execute 'setOrientToAngle' on 'SVGMarkerElement': 1 argument required, but only 0 present..
PASS markerElement.setOrientToAngle(true) threw exception TypeError: Failed to execute 'setOrientToAngle' on 'SVGMarkerElement': parameter 1 is not of type 'SVGAngle'..
PASS markerElement.setOrientToAngle(2) threw exception TypeError: Failed to execute 'setOrientToAngle' on 'SVGMarkerElement': parameter 1 is not of type 'SVGAngle'..
PASS markerElement.setOrientToAngle('aString') threw exception TypeError: Failed to execute 'setOrientToAngle' on 'SVGMarkerElement': parameter 1 is not of type 'SVGAngle'..
PASS markerElement.setOrientToAngle(markerElement) threw exception TypeError: Failed to execute 'setOrientToAngle' on 'SVGMarkerElement': parameter 1 is not of type 'SVGAngle'..
PASS markerElement.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS markerElement.orientAngle.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
